Rename the utils directory to config
